requirements.txt
qrcode[pil]>=7.4.2
Pillow>=10.2.0
pandas>=2.0.0